Move gogrs/cmd -> cmd
