--- cmd/doc.go
@@ -1,27 +1,0 @@
-// Package cmd 一些實作的小工具.
-//
-/*
-主要是來實作套件本身的應用。
-
-Package gogrs-example
-
-簡單範例測試
-
-Package realtime
-
-擷取盤中即時資訊與大盤、上櫃、寶島指數.
-
-Package twsereport
-
-每日收盤後產生符合選股條件的報告
-
-Package twsecache
-
-清除 twse cache.
-
-Package tradingdays-server
-
-提供簡單的日期查詢 API Server.
-
-*/
-package cmd
