+++ doc.go
@@ -21,4 +21,4 @@ Package utils
 套件所需的公用工具
 
 */
-package gogrs
+package main
